feat: [GNS-316] username component
